* trim db user

* added voguepay.com payment method

* fix issue about adding custom link to menu

* fix of # #353

* added start depth setting for bradcrumb

* Additional fixes for Issue #353 (#358)

* other updates

+++ src/Microweber/Traits/ExtendedSave.php
@@ -261,7 +261,7 @@ trait ExtendedSave {
                     $check = $this->app->category_manager->get_items($save_cat_item);
                     if (is_array($check) and !empty($check)){
                         foreach ($check as $item) {
-                            if (!in_array($item['parent_id'], $categories)){
+                            if (!in_array($item['parent_id'], $categories) and !array_search($item['parent_id'], $categories)){
                                 $this->app->category_manager->delete_item($item['id']);
                             }
                         }
